+++ gatsby-site/src/components/full-nav.js
@@ -1,6 +1,8 @@
 import React from "react"
 import Topnav from './Topnav'
 import fullnavstyles from "../styles/fullnav.module.scss"
+import { Link } from "gatsby"
+
 
 const Fullnav = ( props) => {
   return (
@@ -10,16 +12,16 @@ const Fullnav = ( props) => {
         <div className={fullnavstyles.menucontainer}>
             <div className={fullnavstyles.primary}>
                 <ul>
-                    <li>Contact</li>
-                    <li>About</li>
-                    <li>Speaking</li>
+                    <Link className={fullnavstyles.links} to="/contact"><li>Contact</li></Link>
+                    <Link  className={fullnavstyles.links} to="/about" ><li>About</li></Link>
+                    <Link className={fullnavstyles.links} to="/speaking"><li>Speaking</li></Link>
                 </ul>
             </div>
             <div className={fullnavstyles.secondary}>
                 <ul>
-                    <li>Publications</li>
-                    <li>Payment</li>
-                    <li>CV</li>
+                    <Link  className={fullnavstyles.links} to="/publications"><li>Publications</li></Link>
+                    <Link className={fullnavstyles.links} to="/payment"><li>Payment</li></Link>
+                    <Link  className={fullnavstyles.links} to="/cv"><li>CV</li></Link>
                 </ul>
             </div>
         </div>
