+++ gatsby-site/src/styles/fullnav.module.scss
@@ -42,4 +42,6 @@
         
     }
 }
-
+.links{
+    text-decoration: none;
+}
